feat: create context & refactoring
