+++ frontend/package.json
@@ -10,9 +10,12 @@
     "@types/node": "^16.7.13",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
+    "cors": "^2.8.5",
+    "express": "^4.18.1",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-scripts": "5.0.1",
+    "request": "^2.88.2",
     "typescript": "^4.4.2",
     "web-vitals": "^2.1.0"
   },
