+++ frontend/src/App.tsx
@@ -1,11 +1,12 @@
 import { Home } from './Home';
+import { WeatherContextProvider } from './hooks/Weather';
 import './style.css';
 
 function App() {
   return (
-    <div>
+    <WeatherContextProvider>
       <Home />
-    </div>
+    </WeatherContextProvider>
   );
 }
