+++ frontend/src/Home/style.css
@@ -5,75 +5,3 @@
   font-family: 'Roboto', sans-serif;
 }
 
-h2 {
-  font-size: 2.5rem;
-  text-transform: uppercase;
-}
-
-h3 {
-  font-size: 2rem;
-}
-
-p {
-  font-size: 1.5rem;
-  margin-bottom: 20px;
-}
-
-.input-area {
-  display: flex;
-  align-items: center;
-  background: #f0f0f0;
-  padding: 20px;
-  gap: 40px;
-}
-
-.btn {
-  background: transparent;
-  border: none;
-  cursor: pointer;
-  font-size: 1.5rem;
-  color: var(--text);
-}
-
-.btn img {
-  width: 80px;
-}
-
-.weather-input {
-  border: none;
-  background: transparent;
-  font-size: 3rem;
-  flex: 1;
-}
-
-.today {
-  background-color: var(--p1);
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-}
-
-.today-info {
-  padding: 40px 20px 40px 80px;
-}
-
-.today img {
-  width: 420px;
-  padding: 60px;
-}
-
-.tomorrow {
-  background: var(--p2);
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-}
-
-.after-tomorrow {
-  background: var(--p3);
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-}
-
-.content {
-  grid-column: 2;
-  padding: 40px 20px 40px 80px;
-}
