+++ pull-request.txt
@@ -1,3 +1,3 @@
-Your name: ___
-Your Github homepage: ___
-Original challenge URL: http://github.com/hurbcom/challenge-___
+Your name: João Lucas Marcelino
+Your Github homepage: https://github.com/joaomarccelino
+Original challenge URL: https://github.com/hurbcom/challenge-charlie
